file loading updates, debugging, sequelize sync
