+++ packages/spruce-skill-server/models/FileItem.ts
@@ -8,7 +8,8 @@ import { Organization } from './Organization'
 import { User } from './User'
 
 export class FileItem extends SpruceCoreModel<FileItem> {
-	public static readonly timestamps = true
+	// Prevents sequelize from trying to run sync against this model
+	public static readonly doNotSync = true
 	// Scopes
 	public static readonly scopes = {
 		public: {
