+++ packages/spruce-skill-server/models/Job.ts
@@ -8,7 +8,8 @@ import { UserLocation } from './UserLocation'
 import { UserGroup } from './UserGroup'
 
 export class Job extends SpruceCoreModel<Job> {
-	public static readonly timestamps = true
+	// Prevents sequelize from trying to run sync against this model
+	public static readonly doNotSync = true
 	// Scopes
 	public static readonly scopes = {
 		public: {
