+++ packages/spruce-skill-server/models/LocationGroup.ts
@@ -7,7 +7,8 @@ import { Group } from './Group'
 import { Location } from './Location'
 
 export class LocationGroup extends SpruceCoreModel<LocationGroup> {
-	public static readonly timestamps = true
+	// Prevents sequelize from trying to run sync against this model
+	public static readonly doNotSync = true
 	// Scopes
 	public static readonly scopes = {
 		public: {
