+++ packages/spruce-skill-server/models/Organization.ts
@@ -6,7 +6,8 @@ import { SpruceCoreModel, ISpruceCoreSkillModels } from '../types/models'
 import { Location } from './Location'
 
 export class Organization extends SpruceCoreModel<Organization> {
-	public static readonly timestamps = true
+	// Prevents sequelize from trying to run sync against this model
+	public static readonly doNotSync = true
 	// Scopes
 	public static readonly scopes = {
 		public: {
