+++ packages/spruce-skill-server/models/User.ts
@@ -12,7 +12,8 @@ import { Group } from './Group'
 import { UserGroup } from './UserGroup'
 
 export class User extends SpruceCoreModel<User> {
-	public static readonly timestamps = true
+	// Prevents sequelize from trying to run sync against this model
+	public static readonly doNotSync = true
 	// Scopes
 	public static readonly scopes = {
 		public: {
