+++ packages/spruce-skill-server/models/UserOrganization.ts
@@ -7,7 +7,8 @@ import { Organization } from './Organization'
 import { User } from './User'
 
 export class UserOrganization extends SpruceCoreModel<UserOrganization> {
-	public static readonly timestamps = true
+	// Prevents sequelize from trying to run sync against this model
+	public static readonly doNotSync = true
 	// Scopes
 	public static readonly scopes = {
 		public: {
