+++ packages/spruce-skill/package.json
@@ -89,7 +89,6 @@
 		"@semantic-release/changelog": "^3.0.0",
 		"@semantic-release/git": "7.1.0-beta.3",
 		"@sprucelabs/semantic-release": "^2.0.8",
-		"@types/bluebird": "^3.5.27",
 		"babel-jest": "^24.5.0",
 		"chai": "^4.2.0",
 		"conventional-changelog-sprucelabs": "^1.1.0",
