--- packages/spruce-skill/server/models_OLD/Organization.js
@@ -1,34 +1,0 @@
-// 🌲🤖 This is a core model, available if DB_ENABLED=true
-
-// http://docs.sequelizejs.com/manual/tutorial/models-definition.html
-
-const modelName = 'Organization'
-
-module.exports = (sequelize, DataTypes) => {
-	const attributes = {
-		id: {
-			type: DataTypes.UUID,
-			defaultValue: DataTypes.UUIDV4,
-			primaryKey: true
-		},
-		name: {
-			type: DataTypes.STRING
-		}
-	}
-	const options = {
-		timestamps: true
-	}
-	const Organization = sequelize.define(modelName, attributes, options)
-
-	Organization.associate = function(models) {
-		Organization.hasMany(models.Location, { constraints: false })
-	}
-
-	Organization.scopes = {
-		public: {
-			attributes: ['id', 'name']
-		}
-	}
-
-	return Organization
-}
